♻️ refactor: msw 핸들러 함수 분리
