+++ src/api/mswMock/browser.ts
@@ -1,4 +1,10 @@
 import { setupWorker } from 'msw/browser'
-import { handlers } from '@/api/mswMock/handlers'
+import { authHandlers } from '@/api/mswMock/handlers/authHandlers'
+import { myStudiesHandlers } from '@/api/mswMock/handlers/myStudiesHandlers'
+import { studyMakeHandlers } from '@/api/mswMock/handlers/studyMakeHandlers'
 
-export const worker = setupWorker(...handlers)
+export const worker = setupWorker(
+  ...authHandlers,
+  ...myStudiesHandlers,
+  ...studyMakeHandlers
+)
